Ajustes.

+++ src/br/com/reciclagemweb/dao/GenericDAO.java
@@ -8,5 +8,5 @@ public interface GenericDAO<T> {
 	void alterar(T entidade)throws Exception;
 	void deletar(T entidade) throws Exception;
 	List<T> list();
-	T buscar(final long id);
+	T buscar(final int id);
 }
